+++ test/oneOrMore.test.ts
@@ -1,6 +1,6 @@
-import oneOrMore from "./oneOrMore";
-import VerEx from "./verbalexpressions";
-import { startOfLine, endOfLine } from "./constants";
+import oneOrMore from "../src/oneOrMore";
+import VerEx from "../src/verbalexpressions";
+import { startOfLine, endOfLine } from "../src/constants";
 
 describe("oneOrMore", () => {
   it("should export a function", () => {
